Reworked several files so they dont have to save several files each, all is now done on one file each

+++ Pain/pain_superimpose.py
@@ -21,20 +21,15 @@ file_prefix = "normal_eeg_data_"
 file_suffix = ".npy"
 file_index = 0
 
-file_name = file_prefix + str(file_index) + file_suffix
+file_name = "normal_eeg_data.npy"
 file_path = os.path.join(path, file_name)
-while os.path.isfile(file_path):#continue until there is no more file
-    samples = np.load(file_path)
-    print(f"Loaded sample {samples.shape} at {file_name}!")
-
-    data = np.vstack((data, samples))#add to data
-    print(f"Data : {data.shape}")
-    
-    file_index = file_index + 1#iterate to next file
-    file_name =  file_prefix + str(file_index) + file_suffix
-    file_path = os.path.join(path, file_name)#set new file path
-print(f"Found {file_index} files!")
-print(f"All data : {data.shape}")
+if os.path.isfile(file_path):
+    print(f"{file_name} exists!")
+    data = np.load(file_path)
+    print(f"Loaded data : {data.shape}")
+else:
+    print(f"{file_name} does not exist, no data can be loaded!")
+    exit()
 
 samples = []
 existing_recordings = 0#how many recordings already exists
@@ -50,7 +45,8 @@ for sample in range(num_generate_samples):#loop to create
     timing = random.randint(1, recording_time-1)
     intensity = random.randint(1, max_intensity)
     for channel in range(channels):
-        val = data[sample][channel]
+        time_series = data[sample][channel]
+        
     obj = {
         "pain" : pain,
         "timing" : timing,
@@ -64,29 +60,54 @@ print(f"Length of labels {len(labels)}")
 current_dir = pathlib.Path(__file__).parent#get current folder
 path = os.path.join(current_dir, "Datasets", "PainData")#get directory to save recording
 
-file_prefix = "imposed_eeg_data_"
-file_suffix = ".npy"
-file_index = 0
 
-file_name = file_prefix + str(file_index) + file_suffix
+#----------- SAVE EEG DATA IMPOSED WITH PAIN ARTIFACTS --------------
+print(f"---STARTING EEG DATA SAVING---")
+file_name = "imposed_eeg_data.npy"
 file_path = os.path.join(path, file_name)
-while os.path.isfile(file_path):#continue until there is no more file
+if os.path.isfile(file_path):#Does this file already exist ? 
+    #Then we want to add to it
+    print(f"{file_name} exists!")
+    old_data = np.load(file_path)
+    print(f"Loaded data : {old_data.shape} from {file_name}")
+    print(f"Imposed data : {data.shape}")
     
-    file_index = file_index + 1#iterate to next file
-    file_name =  file_prefix + str(file_index) + file_suffix
-    file_path = os.path.join(path, file_name)#set new file path
-#there is no file with this file index
-print(f"Saving at {file_index}")
-np.save(file_path, imposed_data)#save it
-print(f"Saved imposed data {imposed_data.shape} at {file_name}!")
-
-#SAVE JSON FILE WITH LABELS
-json_file_prefix = "imposed_eeg_data_label_"
-json_file_suffix = ".json"
-json_file_name = json_file_prefix + str(file_index) + json_file_suffix
+    new_data = np.vstack((old_data, data))#STACK THE EXISTING DATA 
+    np.save(file_path, new_data) #SAVE THE DATA TO FILE
+    print(f"Saved new data : {new_data.shape} at {file_name}")
+else:
+    print(f"{file_name} does not exists!")
+    saved_data = np.save(file_path, data)
+    print(f"Saved data : {data.shape} at {file_name}")
+    
+#---------- SAVE JSON LABELS ---------------
+print(f"---STARTING JSON SAVING---")
+json_file_name = "imposed_eeg_data_labels.json"
 json_file_path = os.path.join(path, json_file_name)
-with open(json_file_path, "w") as file:
-    json.dump(labels, file)#dump all the files 
-print(f"Saved labels at {json_file_name}!")
+
+json_objs = labels
+if os.path.isfile(json_file_path):#Does the Json file already exist ? 
+    print(f"Json file already exists")
+    
+    loaded_json_objs = []
+    with open(json_file_path, "r") as file:
+        loaded_json_objs = json.load(file)
+        print(f"Loaded json objs : {len(loaded_json_objs)}")
+    #----- WE HAVE LOADED JSON OBJS, NOW WE MUST ADD TO THE END OF IT
+    for label in range(len(loaded_json_objs)):
+        json_objs.append(loaded_json_objs[label])
+    print(f"Combined Json array is {len(json_objs)} long!")
+    with open(json_file_path, "w") as file:
+        json.dump(labels, file)#dump all the files 
+    print(f"Saved {len(json_objs)} labels at {json_file_name}!")
+else:
+    print(f"{json_file_name} does not exist!")
+    
+    with open(json_file_path, "w") as file:
+        json.dump(labels, file)#dump all the files 
+    print(f"Saved {len(json_objs)} labels at {json_file_name}!")
+
+
+
